chore(data-warehouse): move source/schema viewsets to warehouse_sources
The external_data_source, external_data_schema, source_api_versions and
public_source_configs viewsets served warehouse_sources-owned models
(ExternalDataSource/ExternalDataSchema) from the data_warehouse product.
Move them into a new warehouse_sources presentation layer + routes.py so
the views live with their models. Route names are unchanged, so URLs and
generated OpenAPI operation IDs are preserved.

The source-agnostic CI guard now scans both presentation trees so the
guarantee follows the moved views.
